requirements.txt
alembic==1.4.3
appdirs==1.4.4
astroid==2.4.2
click==7.1.2
dnspython==2.0.0
dominate==2.6.0
email-validator==1.1.2
Faker==5.3.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-JWT-Extended==3.25.0
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.1
Flask-WTF==0.14.3
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.1
mypy-extensions==0.4.3
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
regex==2020.11.13
SQLAlchemy==1.3.20
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
